fix: fix soft bug in 4 task (Wed, May 25, 2022, 12:45 PM)

+++ 04-copy-directory/index.js
@@ -6,7 +6,9 @@ async function copy() {
   await fs.mkdir(path.resolve(__dirname, 'files-copy'), {recursive: true});
   const files = await fs.readdir(path.resolve(__dirname, 'files'), {withFileTypes: true});
   files.forEach(async (file) => {
-    await fs.copyFile(path.resolve(__dirname, 'files', file.name), path.resolve(__dirname, 'files-copy', file.name));
+    if (file.isFile()) {
+      await fs.copyFile(path.resolve(__dirname, 'files', file.name), path.resolve(__dirname, 'files-copy', file.name));
+    }
   });
 }
